+++ RELEASE_NOTES.md
@@ -16,6 +16,10 @@ In `v4.0`, this library will become a part of the main library. After that, semv
 * `EmojiCategoryKeyboardMenu` is a new view that lets the user select one of multiple categories.
 * `EmojiKeyboard` is a new view that renders a set of emojis in a lazy grid. The item action is customizable.
 
+### Behavior changes
+
+* `SystemKeyboardButtonBody` now only offsets small caps texts.
+
 ### Deprecations
 
 * `View+keyboardAction(:context:)` has been deprecated.